투표(Vote) 기능 및 항목별 정산 로직 구현

+++ backend/src/main/java/com/jeongchongmu/settlement/dto/DirectSettlementEntry.java
@@ -8,5 +8,5 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class DirectSettlementEntry {
     private Long userId; // 이 사용자가
-    private BigDecimal amount; // 이만큼 부담한다
+    private Long amount; // 이만큼 부담한다
 }
