+++ backend/src/main/java/com/jeongchongmu/settlement/dto/SettlementResponse.java
@@ -17,13 +17,13 @@ public class SettlementResponse {
     private Long expenseId;
     private SettlementMethod method;
     private SettlementStatus status;
-    private BigDecimal totalAmount; // 정산 대상 총 금액 (Expense에서 가져옴)
+    private Long totalAmount; // 정산 대상 총 금액 (Expense에서 가져옴)
 
     // "누가 누구에게 얼마를" 보내야 하는지 상세 내역 리스트
     private List<SettlementDetailDto> details;
 
     // Entity(Settlement)를 Response DTO로 변환하는 정적 메서드 (강력 추천!)
-    public static SettlementResponse from(Settlement settlement, BigDecimal totalAmount) {
+    public static SettlementResponse from(Settlement settlement, Long totalAmount) {
         return SettlementResponse.builder()
                 .settlementId(settlement.getId())
                 .expenseId(settlement.getExpense().getId())
